# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
...
        "/v1/quotes": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Get all quotes belonging to the authenticated user",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "quotes"
                ],
                "summary": "Get all quotes for the authenticated user",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Create a new quote with filament lines and profiles",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "quotes"
                ],
                "summary": "Create a new quote",
                "parameters": [
                    {
                        "description": "Quote data",
                        "name": "quote",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.CreateQuoteRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/dto.QuoteResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/v1/quotes/{id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Get a quote by ID with all its filament lines and profiles",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "quotes"
                ],
                "summary": "Get a quote by ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Quote ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.QuoteResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Update a quote with new data",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "quotes"
                ],
                "summary": "Update a quote",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Quote ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Quote data",
                        "name": "quote",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.UpdateQuoteRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.QuoteResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Delete a quote and all its related data",
                "tags": [
                    "quotes"
                ],
                "summary": "Delete a quote",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Quote ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/v1/quotes/{id}/calculate": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Calculate all costs for a quote based on provided parameters",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "quotes"
                ],
                "summary": "Calculate quote costs",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Quote ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Calculation parameters",
                        "name": "calculation",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.CalculateQuoteRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.CalculationResult"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/v1/quotes/{id}/duplicate": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Create a copy of an existing quote",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "quotes"
                ],
                "summary": "Duplicate a quote",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Quote ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/dto.QuoteResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "dto.CalculateQuoteRequest": {
            "type": "object",
            "required": [
                "print_time_hours",
                "service_type"
            ],
            "properties": {
                "modeler_minutes": {
                    "type": "number",
                    "minimum": 0
                },
                "operator_minutes": {
                    "type": "number",
                    "minimum": 0
                },
                "print_time_hours": {
                    "type": "number",
                    "minimum": 0
                },
                "service_type": {
                    "type": "string",
                    "enum": [
                        "printing_only",
                        "printing_plus",
                        "full_service"
                    ]
                }
            }
        },
        "dto.CalculationResult": {
            "type": "object",
            "properties": {
                "applied_margin": {
                    "type": "number"
                },
                "direct_cost": {
                    "type": "number"
                },
                "energy_cost": {
                    "type": "number"
                },
                "final_price": {
                    "type": "number"
                },
                "labor_cost": {
                    "type": "number"
                },
                "material_cost": {
                    "type": "number"
                },
                "modeler_minutes": {
                    "type": "number"
                },
                "operator_minutes": {
                    "type": "number"
                },
                "print_time_hours": {
                    "type": "number"
                },
                "service_type": {
                    "type": "string"
                },
                "wear_cost": {
                    "type": "number"
                }
            }
        },
        "dto.CostProfileResponse": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "overhead_amount": {
                    "type": "number"
                },
                "quote_id": {
                    "type": "integer"
                },
                "updated_at": {
                    "type": "string"
                },
                "wear_percentage": {
                    "type": "number"
                }
            }
        },
        "dto.CreateCostProfileRequest": {
            "type": "object",
            "properties": {
                "description": {
                    "type": "string"
                },
                "overhead_amount": {
                    "type": "number",
                    "minimum": 0
                },
                "wear_percentage": {
                    "type": "number",
                    "maximum": 100,
                    "minimum": 0
                }
            }
        },
        "dto.CreateEnergyProfileRequest": {
            "type": "object",
            "required": [
                "base_tariff",
                "location",
                "year"
            ],
            "properties": {
                "base_tariff": {
                    "type": "number",
                    "minimum": 0
                },
                "description": {
                    "type": "string"
                },
                "flag_surcharge": {
                    "type": "number",
                    "minimum": 0
                },
                "location": {
                    "type": "string"
                },
                "year": {
                    "type": "integer",
                    "maximum": 2030,
                    "minimum": 2020
                }
            }
        },
        "dto.CreateFilamentLineRequest": {
            "type": "object",
            "required": [
                "weight_grams"
            ],
            "properties": {
                "filament_id": {
                    "description": "Option 1: Automatic snapshot from existing filament",
                    "type": "integer",
                    "minimum": 1
                },
                "filament_snapshot_brand": {
                    "type": "string"
                },
                "filament_snapshot_color": {
                    "type": "string"
                },
                "filament_snapshot_color_hex": {
                    "type": "string"
                },
                "filament_snapshot_material": {
                    "type": "string"
                },
                "filament_snapshot_name": {
                    "description": "Option 2: Manual snapshot data (required if filament_id not provided)",
                    "type": "string"
                },
                "filament_snapshot_price_per_kg": {
                    "type": "number",
                    "minimum": 0
                },
                "filament_snapshot_price_per_meter": {
                    "type": "number",
                    "minimum": 0
                },
                "filament_snapshot_url": {
                    "type": "string"
                },
                "length_meters": {
                    "type": "number",
                    "minimum": 0
                },
                "weight_grams": {
                    "description": "Required fields",
                    "type": "number",
                    "minimum": 0
                }
            }
        },
        "dto.CreateMachineProfileRequest": {
            "type": "object",
            "required": [
                "brand",
                "model",
                "name",
                "watt"
            ],
            "properties": {
                "brand": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "idle_factor": {
                    "type": "number",
                    "maximum": 1,
                    "minimum": 0
                },
                "model": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "url": {
                    "type": "string"
                },
                "watt": {
                    "type": "number",
                    "minimum": 0
                }
            }
        },
        "dto.CreateMarginProfileRequest": {
            "type": "object",
            "properties": {
                "description": {
                    "type": "string"
                },
                "full_service_margin": {
                    "type": "number",
                    "minimum": 0
                },
                "modeler_rate_per_hour": {
                    "type": "number",
                    "minimum": 0
                },
                "operator_rate_per_hour": {
                    "type": "number",
                    "minimum": 0
                },
                "printing_only_margin": {
                    "type": "number",
                    "minimum": 0
                },
                "printing_plus_margin": {
                    "type": "number",
                    "minimum": 0
                }
            }
        },
        "dto.CreateQuoteRequest": {
            "type": "object",
            "required": [
                "filament_lines",
                "title"
            ],
            "properties": {
                "cost_profile": {
                    "$ref": "#/definitions/dto.CreateCostProfileRequest"
                },
                "energy_profile": {
                    "$ref": "#/definitions/dto.CreateEnergyProfileRequest"
                },
                "filament_lines": {
                    "type": "array",
                    "minItems": 1,
                    "items": {
                        "$ref": "#/definitions/dto.CreateFilamentLineRequest"
                    }
                },
                "machine_profile": {
                    "$ref": "#/definitions/dto.CreateMachineProfileRequest"
                },
                "margin_profile": {
                    "$ref": "#/definitions/dto.CreateMarginProfileRequest"
                },
                "notes": {
                    "type": "string"
                },
                "title": {
                    "type": "string",
                    "maxLength": 255,
                    "minLength": 1
                }
            }
        },
...
        "dto.EnergyProfileResponse": {
            "type": "object",
            "properties": {
                "base_tariff": {
                    "type": "number"
                },
                "created_at": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "flag_surcharge": {
                    "type": "number"
                },
                "id": {
                    "type": "integer"
                },
                "location": {
                    "type": "string"
                },
                "quote_id": {
                    "type": "integer"
                },
                "updated_at": {
                    "type": "string"
                },
                "year": {
                    "type": "integer"
                }
            }
        },
        "dto.FilamentLineResponse": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "filament_snapshot_brand": {
                    "type": "string"
                },
                "filament_snapshot_color": {
                    "type": "string"
                },
                "filament_snapshot_color_hex": {
                    "type": "string"
                },
                "filament_snapshot_material": {
                    "type": "string"
                },
                "filament_snapshot_name": {
                    "type": "string"
                },
                "filament_snapshot_price_per_kg": {
                    "type": "number"
                },
                "filament_snapshot_price_per_meter": {
                    "type": "number"
                },
                "filament_snapshot_url": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "length_meters": {
                    "type": "number"
                },
                "quote_id": {
                    "type": "integer"
                },
                "updated_at": {
                    "type": "string"
                },
                "weight_grams": {
                    "type": "number"
                }
            }
        },
...
        "dto.MachineProfileResponse": {
            "type": "object",
            "properties": {
                "brand": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "idle_factor": {
                    "type": "number"
                },
                "model": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "quote_id": {
                    "type": "integer"
                },
                "updated_at": {
                    "type": "string"
                },
                "url": {
                    "type": "string"
                },
                "watt": {
                    "type": "number"
                }
            }
        },
        "dto.MarginProfileResponse": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "full_service_margin": {
                    "type": "number"
                },
                "id": {
                    "type": "integer"
                },
                "modeler_rate_per_hour": {
                    "type": "number"
                },
                "operator_rate_per_hour": {
                    "type": "number"
                },
                "printing_only_margin": {
                    "type": "number"
                },
                "printing_plus_margin": {
                    "type": "number"
                },
                "quote_id": {
                    "type": "integer"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
        "dto.QuoteResponse": {
            "type": "object",
            "properties": {
                "cost_profile": {
                    "$ref": "#/definitions/dto.CostProfileResponse"
                },
                "created_at": {
                    "type": "string"
                },
                "energy_profile": {
                    "$ref": "#/definitions/dto.EnergyProfileResponse"
                },
                "filament_lines": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/dto.FilamentLineResponse"
                    }
                },
                "id": {
                    "type": "integer"
                },
                "machine_profile": {
                    "$ref": "#/definitions/dto.MachineProfileResponse"
                },
                "margin_profile": {
                    "$ref": "#/definitions/dto.MarginProfileResponse"
                },
                "notes": {
                    "type": "string"
                },
                "owner_user_id": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
...
        "dto.UpdateCostProfileRequest": {
            "type": "object",
            "properties": {
                "description": {
                    "type": "string"
                },
                "overhead_amount": {
                    "type": "number",
                    "minimum": 0
                },
                "wear_percentage": {
                    "type": "number",
                    "maximum": 100,
                    "minimum": 0
                }
            }
        },
...
        "dto.UpdateQuoteRequest": {
            "type": "object",
            "required": [
                "filament_lines",
                "title"
            ],
            "properties": {
                "cost_profile": {
                    "$ref": "#/definitions/dto.UpdateCostProfileRequest"
                },
                "energy_profile": {
                    "$ref": "#/definitions/dto.UpdateEnergyProfileRequest"
                },
                "filament_lines": {
                    "type": "array",
                    "minItems": 1,
                    "items": {
                        "$ref": "#/definitions/dto.UpdateFilamentLineRequest"
                    }
                },
                "machine_profile": {
                    "$ref": "#/definitions/dto.UpdateMachineProfileRequest"
                },
                "margin_profile": {
                    "$ref": "#/definitions/dto.UpdateMarginProfileRequest"
                },
                "notes": {
                    "type": "string"
                },
                "title": {
                    "type": "string",
                    "maxLength": 255,
                    "minLength": 1
                }
            }
        },
...
    }
}